fix feature process-interview

+++ app/Services/Interview/InterviewServiceImpl.php
@@ -7,23 +7,26 @@ use App\Models\User;
 use App\Services\DailyCo\DailyCoService;
 use App\Services\GroupQuestion\GroupQuestionService;
 use App\Services\RNews\NewsService;
+use App\Services\RProcess\ProcessService;
 use Carbon\Carbon;
 use Exception;
 use Illuminate\Contracts\Container\BindingResolutionException;
 
 class InterviewServiceImpl implements InterviewService
 {
-    protected $_repository, $newsService, $dailyCoService, $gqService;
+    protected $_repository, $newsService, $dailyCoService, $gqService, $processService;
 
     /**
      * @throws BindingResolutionException
      */
-    public function __construct(NewsService $newsService, DailyCoService $dailyCoService, GroupQuestionService $gqService)
+    public function __construct(NewsService    $newsService, ProcessService $processService,
+                                DailyCoService $dailyCoService, GroupQuestionService $gqService)
     {
         $this->_repository = app()->make(Interview::class);
         $this->newsService = $newsService;
         $this->dailyCoService = $dailyCoService;
         $this->gqService = $gqService;
+        $this->processService = $processService;
     }
 
     /**
@@ -75,6 +78,7 @@ class InterviewServiceImpl implements InterviewService
                     ['news_id', $data['news_id']],
                     ['candidate_id', $data['candidate_id']],
                     ['company_id', $data['company_id']],
+                    ['process_id', $data['process_id'] ?? null],
                 ])->first();
 
                 if ($interview) {
@@ -98,6 +102,31 @@ class InterviewServiceImpl implements InterviewService
                     $group = $this->gqService->createRandom($user, $data, null);
                     if ($group) {
                         $data = array_merge($data, ['gq_test_id' => $group->id]);
+                    }
+                }
+
+                if (isset($data['process_id'])) {
+                    $process = $this->processService->findById($user, $data['process_id']);
+                    if ($process) {
+                        $prevStep = $this->_repository->where([
+                            ['news_id', $data['news_id']],
+                            ['candidate_id', $data['candidate_id']],
+                            ['company_id', $data['company_id']],
+                        ])->where('process_id', '<', $data['process_id'])
+                            ->orderBy('process_id', 'desc')->first();
+                        $nextStep = $this->_repository->where([
+                            ['news_id', $data['news_id']],
+                            ['candidate_id', $data['candidate_id']],
+                            ['company_id', $data['company_id']],
+                        ])->where('process_id', '>', $data['process_id'])
+                            ->orderBy('process_id', 'asc')->first();
+
+                        if (empty($nextStep) && (empty($prevStep) || $prevStep->is_success)) {
+                            $data['process_id'] = $process->id;
+                        } else {
+                            return false;
+                        }
+
                     }
                 }
 
@@ -142,13 +171,17 @@ class InterviewServiceImpl implements InterviewService
     {
         try {
             if ($user->candidate_id) {
-                $interview = $this->_repository->where([
-                    ['news_id', $news_id],
-                    ['candidate_id', $user->candidate_id],
-                ])->first();
+                $news = $this->newsService->findById($user, $news_id);
+                if ($news) {
+                    $interview = $this->_repository->join('recruitment_news', 'interview.news_id', '=', 'recruitment_news.id')
+                        ->where([
+                            ['project_id', $news->project_id],
+                            ['candidate_id', $user->candidate_id],
+                        ])->selectRaw("interview.*")->first();
 
-                if ($interview) {
-                    return $interview;
+                    if ($interview) {
+                        return $interview;
+                    }
                 }
             }
 
@@ -229,18 +262,34 @@ class InterviewServiceImpl implements InterviewService
                     $data['result'] = ($result);
                 }
 
+                if (isset($data['process_id']) && !$interview->process_id) {
+                    $process = $this->processService->findById($user, $data['process_id']);
+                    if ($process) {
+                        $data['process_id'] = $process->id;
+                    }
+                } else {
+                    unset($data['process_id']);
+                }
+
+                if (isset($data['gq_interview_id'])) {
+                    $interviewQs = $this->gqService->findById($user, $data['gq_interview_id']);
+                    if ($interviewQs) {
+                        $data['gq_interview_id'] = $interviewQs->id;
+                    }
+                }
+
                 if ($interview->room && (!empty($data['result']) || !empty($data['record']))) {
                     $array = explode('/', $interview->room);
                     $room_name = end($array);
                     $this->dailyCoService->delete($user, $room_name);
                     $data['room'] = null;
                 } else if (empty($data['result']) && empty($data['record'])) {
-                    if ($data['form'] === 'Online' && !$interview->room) {
+                    if (isset($data['form']) && $data['form'] === 'Online' && !$interview->room) {
                         $room = $this->createRoom($user);
                         if ($room) {
                             $data['room'] = $room;
                         }
-                    } else if ($data['form'] === 'Offline' && $interview->room) {
+                    } else if (isset($data['form']) && $data['form'] === 'Offline' && $interview->room) {
                         $array = explode('/', $interview->room);
                         $room_name = end($array);
                         $this->dailyCoService->delete($user, $room_name);
@@ -275,7 +324,14 @@ class InterviewServiceImpl implements InterviewService
                     $room_name = end($array);
                     $this->dailyCoService->delete($user, $room_name);
                 }
-                return $interview->delete();
+
+                $this->_repository->where([
+                    ['candidate_id', $interview->candidate_id],
+                    ['company_id', $interview->company_id],
+                    ['news_id', $interview->news_id],
+                ])->delete();
+
+                return true;
             }
 
             return false;
@@ -288,22 +344,58 @@ class InterviewServiceImpl implements InterviewService
     /**
      * @inheritDoc
      */
-    public function showAllByNews(User $user, int $news_id, int $per_page = 8)
+    public function showAllByNews(User $user, $news_id, $process_id)
     {
         try {
-            $news = $this->newsService->findById($user, $news_id);
+            $data = [];
+            $news = null;
+            if (!empty($news_id)) {
+                $news = $this->newsService->findById($user, $news_id);
 
-            if ($news && $news->company->id == $user->company_id) {
-                $interviews = $this->_repository->where([
-                    ['company_id', $user->company_id],
-                    ['news_id', $news_id]
-                ])
-                    ->orderBy('created_at', 'desc')
-                    ->get();
+                if ($news && $news->company->id == $user->company_id) {
+                    $data = $this->_repository->where([
+                        ['company_id', $user->company_id],
+                        ['news_id', $news_id]
+                    ])->orderBy('created_at', 'desc')
+                        ->get();
+                }
+            } else if (!empty($process_id)) {
+                $process = $this->processService->findById($user, $process_id);
 
+                if ($process && $process->project && $process->project->company_id == $user->company_id) {
+                    $interviews = $this->_repository->join('recruitment_news', 'interview.news_id', '=', 'recruitment_news.id')
+                        ->where([
+                            ['project_id', $process->project_id],
+                            ['interview.company_id', $user->company_id]
+                        ])
+                        ->where(function ($query) use ($process) {
+                            $query->where('process_id', '<=', $process->id)
+                                ->orWhere('process_id', null);
+                        })->selectRaw("interview.*")->orderBy('created_at', 'desc')->get();
+
+                    foreach ($interviews as $interview) {
+                        if ($interview->process_id == $process_id || empty($interview->process_id)) {
+                            $data[] = $interview;
+                        } else {
+                            $nextStep = $this->_repository->where([
+                                ['news_id', $interview->news_id],
+                                ['candidate_id', $interview->candidate_id],
+                                ['company_id', $interview->company_id],
+                            ])->where('process_id', '>', $interview->process_id)
+                                ->orderBy('process_id', 'asc')->first();
+
+                            if ($interview->is_success && empty($nextStep)) {
+                                $data[] = $interview;
+                            }
+                        }
+                    }
+                }
+            }
+
+            if ($data) {
                 return [
-                    'data' => $interviews,
-                    'news' => $news,
+                    'data' => $data,
+                    'news' => $news
                 ];
             }
